requirements.txt
streamlit
opencv-python
pandas
ultralytics
transformers
torch
pillow
numpy
